+++ example/form/components/progress/index.tsx
@@ -1,7 +1,17 @@
+/** @jsxImportSource @emotion/react */
 import { useContext } from '../../../../src/context/index.js';
+import * as styles from './styles.js';
 
 export default function Progress() {
     const form = useContext();
 
-    return <pre>{JSON.stringify(form.progress, null, 2)}</pre>;
+    return (
+        <div css={styles.container}>
+            {form.progress.map((item) => (
+                <div key={item.step} css={styles.item(item.current)}>
+                    {item.step}
+                </div>
+            ))}
+        </div>
+    );
 }